refactor(models): extract useModelsData hook from models-page (phase 1)
First phase of the models-page.tsx split — the canonical extract-custom-hook
refactor. ModelsPage carried ~60 interlocking pieces of state, seven
fetchers, four effects, every action handler, and all derived values inline.

lib-side `use-models-data.ts` exports `useModelsData()` holding all of it:
state, fetchers (config/available/refresh/aliases/spend/budget/routing), the
mount + spend-tab + routing-tab + stale-auto-refresh effects (the last keeps
its deliberate partial dependency array + eslint-disable so the background
refresh can't double-fire), the action handlers (agent/defaults/alias/routing/
budget), and the derived values (modelOptions, modelsReady, availableProviders,
effective*, fallbackCandidates, displayRouting, hasPending, defaultsDirty). The
local Routing*/Budget*/Spend* interfaces move with it.

models-page.tsx shell destructures the one returned object, so the entire
return(...) JSX is byte-identical. Behavior-preserving — same hook-call order,
same effects. The ModelsPage export stays put (client.tsx slot + the test
import unchanged). 1,205 -> 804 shell + 458 hook.

This is the foundation for phase 2 (per-tab components consuming the hook);
the appendix's redesigns (sentinel cleanup, postJson, batch saveAll, SDK
extractions of formatAge/RuntimeRestartBanner/TableSkeleton) stay deferred.

Verified: typecheck + eslint + models-page.test (10/0) + models + components
suites (341/0) + full suite (5124/0) + binary build + isolated-home boot smoke
(Models plugin loads; rebuilt client.js → 200). The /config 500 in the smoke
is environmental (runtime adapter absent without OpenClaw), not a regression —
this change is client-only and doesn't touch that route.
